+++ debugd/internal/debugd/metadata/cloudprovider/cloudprovider.go
@@ -12,7 +12,6 @@ import (
 	"net"
 
 	"github.com/edgelesssys/constellation/v2/internal/cloud/metadata"
-	"github.com/edgelesssys/constellation/v2/internal/deploy/ssh"
 	"github.com/edgelesssys/constellation/v2/internal/role"
 )
 
@@ -25,7 +24,7 @@ type providerMetadata interface {
 	GetLoadBalancerEndpoint(ctx context.Context) (string, error)
 }
 
-// Fetcher checks the metadata service to search for instances that were set up for debugging and cloud provider specific SSH keys.
+// Fetcher checks the metadata service to search for instances that were set up for debugging.
 type Fetcher struct {
 	metaAPI providerMetadata
 }
@@ -91,21 +90,4 @@ func (f *Fetcher) DiscoverLoadbalancerIP(ctx context.Context) (string, error) {
 	}
 
 	return lbIP, nil
-}
-
-// FetchSSHKeys will query the metadata of the current instance and deploys any SSH keys found.
-func (f *Fetcher) FetchSSHKeys(ctx context.Context) ([]ssh.UserKey, error) {
-	self, err := f.metaAPI.Self(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("retrieving ssh keys from cloud provider metadata: %w", err)
-	}
-
-	keys := []ssh.UserKey{}
-	for username, userKeys := range self.SSHKeys {
-		for _, keyValue := range userKeys {
-			keys = append(keys, ssh.UserKey{Username: username, PublicKey: keyValue})
-		}
-	}
-
-	return keys, nil
 }