+++ debugd/service/debugd_grpc.pb.go
@@ -22,10 +22,8 @@ const _ = grpc.SupportPackageIsVersion7
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type DebugdClient interface {
-	UploadAuthorizedKeys(ctx context.Context, in *UploadAuthorizedKeysRequest, opts ...grpc.CallOption) (*UploadAuthorizedKeysResponse, error)
 	UploadBootstrapper(ctx context.Context, opts ...grpc.CallOption) (Debugd_UploadBootstrapperClient, error)
 	DownloadBootstrapper(ctx context.Context, in *DownloadBootstrapperRequest, opts ...grpc.CallOption) (Debugd_DownloadBootstrapperClient, error)
-	DownloadAuthorizedKeys(ctx context.Context, in *DownloadAuthorizedKeysRequest, opts ...grpc.CallOption) (*DownloadAuthorizedKeysResponse, error)
 	UploadSystemServiceUnits(ctx context.Context, in *UploadSystemdServiceUnitsRequest, opts ...grpc.CallOption) (*UploadSystemdServiceUnitsResponse, error)
 }
 
@@ -35,15 +33,6 @@ type debugdClient struct {
 
 func NewDebugdClient(cc grpc.ClientConnInterface) DebugdClient {
 	return &debugdClient{cc}
-}
-
-func (c *debugdClient) UploadAuthorizedKeys(ctx context.Context, in *UploadAuthorizedKeysRequest, opts ...grpc.CallOption) (*UploadAuthorizedKeysResponse, error) {
-	out := new(UploadAuthorizedKeysResponse)
-	err := c.cc.Invoke(ctx, "/debugd.Debugd/UploadAuthorizedKeys", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
 }
 
 func (c *debugdClient) UploadBootstrapper(ctx context.Context, opts ...grpc.CallOption) (Debugd_UploadBootstrapperClient, error) {
@@ -112,15 +101,6 @@ func (x *debugdDownloadBootstrapperClient) Recv() (*Chunk, error) {
 	return m, nil
 }
 
-func (c *debugdClient) DownloadAuthorizedKeys(ctx context.Context, in *DownloadAuthorizedKeysRequest, opts ...grpc.CallOption) (*DownloadAuthorizedKeysResponse, error) {
-	out := new(DownloadAuthorizedKeysResponse)
-	err := c.cc.Invoke(ctx, "/debugd.Debugd/DownloadAuthorizedKeys", in, out, opts...)
-	if err != nil {
-		return nil, err
-	}
-	return out, nil
-}
-
 func (c *debugdClient) UploadSystemServiceUnits(ctx context.Context, in *UploadSystemdServiceUnitsRequest, opts ...grpc.CallOption) (*UploadSystemdServiceUnitsResponse, error) {
 	out := new(UploadSystemdServiceUnitsResponse)
 	err := c.cc.Invoke(ctx, "/debugd.Debugd/UploadSystemServiceUnits", in, out, opts...)
@@ -134,10 +114,8 @@ func (c *debugdClient) UploadSystemServiceUnits(ctx context.Context, in *UploadS
 // All implementations must embed UnimplementedDebugdServer
 // for forward compatibility
 type DebugdServer interface {
-	UploadAuthorizedKeys(context.Context, *UploadAuthorizedKeysRequest) (*UploadAuthorizedKeysResponse, error)
 	UploadBootstrapper(Debugd_UploadBootstrapperServer) error
 	DownloadBootstrapper(*DownloadBootstrapperRequest, Debugd_DownloadBootstrapperServer) error
-	DownloadAuthorizedKeys(context.Context, *DownloadAuthorizedKeysRequest) (*DownloadAuthorizedKeysResponse, error)
 	UploadSystemServiceUnits(context.Context, *UploadSystemdServiceUnitsRequest) (*UploadSystemdServiceUnitsResponse, error)
 	mustEmbedUnimplementedDebugdServer()
 }
@@ -146,17 +124,11 @@ type DebugdServer interface {
 type UnimplementedDebugdServer struct {
 }
 
-func (UnimplementedDebugdServer) UploadAuthorizedKeys(context.Context, *UploadAuthorizedKeysRequest) (*UploadAuthorizedKeysResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method UploadAuthorizedKeys not implemented")
-}
 func (UnimplementedDebugdServer) UploadBootstrapper(Debugd_UploadBootstrapperServer) error {
 	return status.Errorf(codes.Unimplemented, "method UploadBootstrapper not implemented")
 }
 func (UnimplementedDebugdServer) DownloadBootstrapper(*DownloadBootstrapperRequest, Debugd_DownloadBootstrapperServer) error {
 	return status.Errorf(codes.Unimplemented, "method DownloadBootstrapper not implemented")
-}
-func (UnimplementedDebugdServer) DownloadAuthorizedKeys(context.Context, *DownloadAuthorizedKeysRequest) (*DownloadAuthorizedKeysResponse, error) {
-	return nil, status.Errorf(codes.Unimplemented, "method DownloadAuthorizedKeys not implemented")
 }
 func (UnimplementedDebugdServer) UploadSystemServiceUnits(context.Context, *UploadSystemdServiceUnitsRequest) (*UploadSystemdServiceUnitsResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method UploadSystemServiceUnits not implemented")
@@ -172,24 +144,6 @@ type UnsafeDebugdServer interface {
 
 func RegisterDebugdServer(s grpc.ServiceRegistrar, srv DebugdServer) {
 	s.RegisterService(&Debugd_ServiceDesc, srv)
-}
-
-func _Debugd_UploadAuthorizedKeys_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(UploadAuthorizedKeysRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(DebugdServer).UploadAuthorizedKeys(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/debugd.Debugd/UploadAuthorizedKeys",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(DebugdServer).UploadAuthorizedKeys(ctx, req.(*UploadAuthorizedKeysRequest))
-	}
-	return interceptor(ctx, in, info, handler)
 }
 
 func _Debugd_UploadBootstrapper_Handler(srv interface{}, stream grpc.ServerStream) error {
@@ -239,24 +193,6 @@ func (x *debugdDownloadBootstrapperServer) Send(m *Chunk) error {
 	return x.ServerStream.SendMsg(m)
 }
 
-func _Debugd_DownloadAuthorizedKeys_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
-	in := new(DownloadAuthorizedKeysRequest)
-	if err := dec(in); err != nil {
-		return nil, err
-	}
-	if interceptor == nil {
-		return srv.(DebugdServer).DownloadAuthorizedKeys(ctx, in)
-	}
-	info := &grpc.UnaryServerInfo{
-		Server:     srv,
-		FullMethod: "/debugd.Debugd/DownloadAuthorizedKeys",
-	}
-	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
-		return srv.(DebugdServer).DownloadAuthorizedKeys(ctx, req.(*DownloadAuthorizedKeysRequest))
-	}
-	return interceptor(ctx, in, info, handler)
-}
-
 func _Debugd_UploadSystemServiceUnits_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(UploadSystemdServiceUnitsRequest)
 	if err := dec(in); err != nil {
@@ -282,14 +218,6 @@ var Debugd_ServiceDesc = grpc.ServiceDesc{
 	ServiceName: "debugd.Debugd",
 	HandlerType: (*DebugdServer)(nil),
 	Methods: []grpc.MethodDesc{
-		{
-			MethodName: "UploadAuthorizedKeys",
-			Handler:    _Debugd_UploadAuthorizedKeys_Handler,
-		},
-		{
-			MethodName: "DownloadAuthorizedKeys",
-			Handler:    _Debugd_DownloadAuthorizedKeys_Handler,
-		},
 		{
 			MethodName: "UploadSystemServiceUnits",
 			Handler:    _Debugd_UploadSystemServiceUnits_Handler,